Allow listener log to be specified as an argument
Add initial support for processing of commandline arguments, starting
with -f|--file to indicate the name of the listener logfile to be
processed. When this option is specified it's no longer necessary to run
and parse the output of the `lsnrctl status` command, which means that
the log can be processed on a machine other than the database server.

Example usage:

   analyze_lsnr_log.py -f listener.log
   analyze_lsnr_log.py -f listener_test.log
   analyze_lsnr_log.py --file /path/to/listener.log

Due to use of the OptionParser module, the -h|--help option is also
recognized and will provide an example of the available options and
syntax.