Handle case where PROGRAM entry is unspecified
In some cases, program name can be empty in the listener log. The
CONNECT_DATA clause will be similar to:

   (CONNECT_DATA=(CID=(PROGRAM=)(HOST=xxx.example.com)...

This was causing the script to mis-parse the program name, erroneously
identifying it as ")(HOST" rather than a missing value. The solution
chosen was to tag these items as "*UNSPECIFIED*" in the output.